+++ Database/BroadcastDatabase/Scripts/0026-Data.UpdateEventsFrom.sql
@@ -1,0 +1,5 @@
+﻿update EmailTemplate 
+set	[from] = 'events@kandobay.com.au'
+where DocType like 'Event%'
+and Brand = 'KandoBay'
+